feat: Extract a first class XML encoder/parser
