+++ spec/faraday/xml/response_spec.rb
@@ -116,7 +116,7 @@ RSpec.describe Faraday::XML::Response, type: :response do # rubocop:disable RSpe
 
     it 'passes relevant options to XML parse' do # rubocop:disable RSpec/ExampleLength, RSpec/MultipleExpectations
       allow(middleware).to receive(:test_parser)
-      expect(middleware.parser).to receive(:call) # rubocop:disable RSpec/MessageSpies, RSpec/StubbedMock
+      expect(middleware.parser).to receive(:parse) # rubocop:disable RSpec/MessageSpies, RSpec/StubbedMock
         .with(body, options[:parser_options] || {})
         .and_return(result)
